requirements.txt
wheel
pre-commit
gdown
numpy
scipy
scikit-learn
Cython
matplotlib
pandas
seaborn
shapely
pyyaml
ipympl
open3d>=0.17.0
pytest